Restructure list page and redirect unauthorized users to index
